refactor by abstracting away events validation | db logic
