+++ functions/api/src/api/rules/event.rules.ts
@@ -1,0 +1,3 @@
+import { param } from 'express-validator'
+
+export const eventIdParamRule = param('eventId').notEmpty().isString()
