+++ functions/api/src/dtos/event.data.ts
@@ -1,0 +1,1 @@
+export type EventData = Record<string, unknown> & { id: string }
